+++ orangecontrib/network/widgets/tests/test_OWNxFile.py
@@ -1,14 +1,11 @@
-import os
 import unittest
 from unittest.mock import patch, Mock
 
-from Orange.widgets.tests.base import WidgetTest
-
 from orangecontrib.network.widgets.OWNxFile import OWNxFile
-import orangecontrib.network
+from orangecontrib.network.widgets.tests.utils import NetworkTest
 
 
-class TestOWNxFile(WidgetTest):
+class TestOWNxFile(NetworkTest):
     def setUp(self):
         self.widget = self.create_widget(OWNxFile)  # type: OWNxFile
 
@@ -17,9 +14,7 @@ class TestOWNxFile(WidgetTest):
                    Mock(side_effect=OSError)):
             self.widget.open_net_file("foo.net")
         self.assertTrue(self.widget.Error.io_error.is_shown())
-        filename = os.path.join(
-            os.path.split(orangecontrib.network.__file__)[0],
-            "networks/leu_by_genesets.net")
+        filename = self._get_filename("leu_by_genesets.net")
         self.widget.open_net_file(filename)
         self.assertFalse(self.widget.Error.io_error.is_shown())
