+++ orangecontrib/network/widgets/tests/test_OWNxGenerator.py
@@ -1,10 +1,8 @@
-from Orange.widgets.tests.base import WidgetTest
-
 from orangecontrib.network.widgets.OWNxGenerator import OWNxGenerator
+from orangecontrib.network.widgets.tests.utils import NetworkTest
 
 
-class TestOWNxGenerator(WidgetTest):
+class TestOWNxGenerator(NetworkTest):
 
     def setUp(self):
         self.widget = self.create_widget(OWNxGenerator)  # type: OWNxGenerator
-
